+++ intellectuals-webapp/package.json
@@ -21,6 +21,7 @@
     "react-bootstrap": "^2.9.1",
     "react-dom": "^18.2.0",
     "react-icons": "^4.12.0",
+    "react-p5-wrapper": "^4.1.2",
     "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
     "react-three-fiber": "^6.0.13",
