+++ intellectuals-webapp/src/css/Timeline.css
@@ -1,13 +1,16 @@
 /* Base styles for the timeline container */
 /* Base styles for the timeline container */
 .timeline-container {
+    margin-top: 100vh;
+    z-index: 4;
     position: relative;
-    padding-top: 2rem;
+    padding-top: 18vh;
     padding-bottom: 2rem;
     background-color: black;
     display: flex;
     flex-direction: column;
     align-items: center;
+    border-top: 10px solid #E63946;
 }
 
 /* Styles for the timeline vertical line */
